♻️ 편지관련 서버통신 코드를 Model의 LetterManager 안으로 리팩토링
추후의 편지 관련 요청들도 LetterManager 파일에 추가 예정
LetterManagerDelegate 프로토콜 추가
